feat: Update Home component to fetch and display Pokemon data
The code changes in the Home component include updating the fetch request to the PokeAPI to retrieve a list of Pokemon data. The fetched data is then stored in the  state variable. Additionally, the  state variable is used to indicate the loading status of the data. This commit enhances the project by fetching and displaying Pokemon data on the Home page.

Note: The recent user commits and repository commits do not provide any additional relevant information for the commit message.